Module pallet_fees::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

GenesisConfig

Genesis configuration

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

AccountFees

Account fees accumulated by eras

ActiveEra

The active era information, it holds index and start.

CurrentSession

The current session of the era.

EraTotalFees

The total fees for the era. If total hasn’t been set or has been removed then 0 stake is returned.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

SessionTotalFees

The total fees for the session. If total hasn’t been set or has been removed then 0 stake is returned.

StoredSessions

Map from all stored sessions.