Trait pallet_fees::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type UnixTime: UnixTime;
    type FeesPalletId: Get<PalletId>;
    type SessionsPerEra: Get<SessionIndex>;
    type SessionsArchive: Get<SessionIndex>;
    type BlocksPerSession: Get<Self::BlockNumber>;
    type WeightInfo: WeightInfo;
    type Security: SecurityExt<Self::AccountId, Self::BlockNumber>;
    type Staking: StakingExt<Self::AccountId>;
    type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
    type ForceOrigin: EnsureOrigin<Self::Origin>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Events

type UnixTime: UnixTime

Time used for computing era duration.

It is guaranteed to start being called from the first on_finalize. Thus value at genesis is not used.

type FeesPalletId: Get<PalletId>

Pallet ID

type SessionsPerEra: Get<SessionIndex>

Number of sessions per era

type SessionsArchive: Get<SessionIndex>

Number of sessions to keep in archive

type BlocksPerSession: Get<Self::BlockNumber>

Number of block per session

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Security: SecurityExt<Self::AccountId, Self::BlockNumber>

Security traits

type Staking: StakingExt<Self::AccountId>

Tidefi stake traits

type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Tidechain currency wrapper

type ForceOrigin: EnsureOrigin<Self::Origin>

The origin which may forcibly update the fee and distribution percentage